Spelling, plus altered ports
