+++ src/Proxy/default.conf
@@ -6,21 +6,21 @@ server {
     resolver 127.0.0.11 ipv6=off valid=1s;
     set $upstream products;
 
-    proxy_pass http://$upstream:5000/$1;
+    proxy_pass http://$upstream:5001/$1;
   }
 
   location ~ /ratings/(.*) {
     resolver 127.0.0.11 ipv6=off valid=1s;
     set $upstream ratings;
 
-    proxy_pass http://$upstream:5000/$1;
+    proxy_pass http://$upstream:5002/$1;
   }
 
-  location ~ /recommandations/(.*) {
+  location ~ /recommendations/(.*) {
     resolver 127.0.0.11 ipv6=off valid=1s;
-    set $upstream recommandations;
+    set $upstream recommendations;
 
-    proxy_pass http://$upstream:5000/$1;
+    proxy_pass http://$upstream:5003/$1;
   }
 
   location ~ /(.*) {
