--- src/RecommandationsApi/.dockerignore
@@ -1,2 +1,0 @@
-bin/
-obj/
