+++ src/RecommendationsApi/.dockerignore
@@ -1,0 +1,2 @@
+bin/
+obj/
