+++ src/ShopFront/Models/Settings.cs
@@ -5,7 +5,7 @@ namespace ShopFront.Models
     public class Settings
     {
         public string ProductsApiUrl { get; set; }
-        public string RecommandationsApiUrl { get; set; }
+        public string RecommendationsApiUrl { get; set; }
         public string RatingsApiUrl { get; set; }
     }
 }
