+++ src/ShopFront/Startup.cs
@@ -35,7 +35,7 @@ namespace ShopFront
             services.Configure<ShopFront.Models.Settings>(settings => 
             {
                 settings.ProductsApiUrl = Configuration["SHOP_PRODUCTS_API_URL"];
-                settings.RecommandationsApiUrl = Configuration["SHOP_RECOMMANDATIONS_API_URL"];
+                settings.RecommendationsApiUrl = Configuration["SHOP_RECOMMANDATIONS_API_URL"];
                 settings.RatingsApiUrl = Configuration["SHOP_RATINGS_API_URL"];
             });
         }
